Added more fixture cutting and class injection.
[Details]
Added advanced pytest fixture cutting for cut. Cut now looks for
fixtures in multiple pytest marks.
Also added code to try really hard to get a class to work if it isn't
already in the test. This will probably not work in most cases though
because if the class is not there, then it is likely that the imports
aren't either.

[File's changed]

	modified:   cut
	modified:   splice