Made multi-fault version of bugsinpy-checkout.
[Details]
Made the multi-fault version of the checkout functionality for bugsinpy.
This checks-out the version given, and then transplants all the
identified relevant tests.

[Note]
Also fixed some issues in the bugsinpy-to-tcm script.

[Bugs]
Pytest seems to have irregular output using different versions. Fixed
the bugsinpy-to-tcm script to deal with these, but they might crop up
again.

[File's changed]

	new file:   framework/bin/bugsinpy-multi-checkout
	modified:   framework/bin/bugsinpy-to-tcm