Added backtracking and commit extracting scripts.
[Details]
Created the extract_shas.sh script, which extracts the necessary commit
SHAs (hashes) needed to do the bug backtracking using the bug_backtracker
(https://github.com/DCallaz/bug-backtracker).
Also added the backtracking.py script to read the json file format of
the bug_backtracker to retrieve the faulty lines for a bug in a
different version.

[File's changed]

	new file:   backtrack.py
	new file:   extract_shas.sh