Refactored bugsinpy-cut, recursive func handling.
[Details]
Refactored the code in bugsinpy-cut so that the dependency collector is
a separate function and can be called recursively for each function
added. This allows dependencies of a function dependency to be added as
well, which should give a fully complete subsystem of the test that is
always able to run. Also fixed the handling of global variables so that
dependencies in them are found as well.

[File's changed]

	modified:   bugsinpy-cut