Fixed bugsinpy-test for single test running.
[Details]
The last fix for the single test running was not sufficient for tests
using tox. For this reason, the single test running was completely
rewritten. It now takes the commands given in bugsinpy_run_tests.sh for
the current version, and replaces the test in this command with the test
that is given. This produces much more stable results as these commands
are hand-written for each version.

[File's changed]

	modified:   bugsinpy-test