Added projects directory from BugsInPy.
[Details]
Due to a significant amount of changes to the projects in the projects
directory of the underlying BugsInPy dataset, it is now necessary to
include this directory in this multifault version. This makes this
multifault version a complete standalone version, which does not require
any files from the original BugsInPy dataset.

[Note]

[Bugs]

[File's changed]

	new file:   projects/*