Added a dos2unix conversion after compiling.
[Details]
On some of the projects (matplotlib for example), the compile script
added files in the env directory that were created in DOS format. This
then would produce a large amount of warnings when running certain
BugsInPy commands. Therefore a dos2unix conversion was added at the end
of bugsinpy-compile to rectify the format of these files.

[File's changed]

	modified:   framework/bin/bugsinpy-compile